// sdk/frontend/src/protocol/types.ts
export interface I18nConfiguration {
  defaultLocale: string
  locales: string[]
  messages: Record<string, Record<string, string>>
}